+++ src/test/java/kr/hhplus/be/ecommerce/application/balance/BalanceFacadeUnitTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.ecommerce.application.balance;
 
-import kr.hhplus.be.ecommerce.MockTestSupport;
+import kr.hhplus.be.ecommerce.support.MockTestSupport;
 import kr.hhplus.be.ecommerce.domain.balance.BalanceInfo;
 import kr.hhplus.be.ecommerce.domain.balance.BalanceService;
 import kr.hhplus.be.ecommerce.domain.user.UserService;
