+++ src/test/java/kr/hhplus/be/ecommerce/application/order/OrderFacadeIntegrationTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.ecommerce.application.order;
 
-import kr.hhplus.be.ecommerce.IntegrationTestSupport;
+import kr.hhplus.be.ecommerce.support.IntegrationTestSupport;
 import kr.hhplus.be.ecommerce.domain.balance.Balance;
 import kr.hhplus.be.ecommerce.domain.balance.BalanceRepository;
 import kr.hhplus.be.ecommerce.domain.coupon.Coupon;
