+++ src/test/java/kr/hhplus/be/ecommerce/application/order/OrderFacadeUnitTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.ecommerce.application.order;
 
-import kr.hhplus.be.ecommerce.MockTestSupport;
+import kr.hhplus.be.ecommerce.support.MockTestSupport;
 import kr.hhplus.be.ecommerce.domain.balance.BalanceService;
 import kr.hhplus.be.ecommerce.domain.coupon.CouponInfo;
 import kr.hhplus.be.ecommerce.domain.coupon.CouponService;
