+++ src/test/java/kr/hhplus/be/ecommerce/application/user/UserCouponFacadeUnitTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.ecommerce.application.user;
 
-import kr.hhplus.be.ecommerce.MockTestSupport;
+import kr.hhplus.be.ecommerce.support.MockTestSupport;
 import kr.hhplus.be.ecommerce.domain.coupon.CouponInfo;
 import kr.hhplus.be.ecommerce.domain.coupon.CouponService;
 import kr.hhplus.be.ecommerce.domain.user.UserCouponInfo;
