+++ src/test/java/kr/hhplus/be/ecommerce/docs/product/ProductControllerDocsTest.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.ecommerce.docs.product;
 import kr.hhplus.be.ecommerce.application.product.ProductFacade;
 import kr.hhplus.be.ecommerce.application.product.ProductResult;
 import kr.hhplus.be.ecommerce.interfaces.product.ProductController;
-import kr.hhplus.be.ecommerce.docs.RestDocsSupport;
+import kr.hhplus.be.ecommerce.support.RestDocsSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.restdocs.payload.JsonFieldType;
