+++ src/test/java/kr/hhplus/be/ecommerce/domain/order/OrderServiceUnitTest.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.ecommerce.domain.order;
 
-import kr.hhplus.be.ecommerce.MockTestSupport;
+import kr.hhplus.be.ecommerce.support.MockTestSupport;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
