+++ src/test/java/kr/hhplus/be/ecommerce/support/IntegrationTestSupport.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.ecommerce;
+package kr.hhplus.be.ecommerce.support;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
