+++ src/test/java/kr/hhplus/be/ecommerce/support/MockTestSupport.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.ecommerce;
+package kr.hhplus.be.ecommerce.support;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
